Added StateManagement, And Remeber TextFieldInput during rotational building.

+++ app/src/main/java/com/cc221012_cc221016/stash/ui/views/composables/ShowEntryView.kt
@@ -56,7 +56,7 @@ import kotlinx.coroutines.launch
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ShowEntryView(
-    entryId: Long,
+    entryId: Int,
     onBack: () -> Unit,
     mainViewModel: MainViewModel,
     onEditEntry: (Entries) -> Unit,
